gestion_usuarios/requirements.txt
Django==5.1.1
annotated-types==0.7.0
asgiref==3.8.1
django-ninja==1.3.0
pydantic==2.9.1
pydantic_core==2.23.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1